ml-backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
opencv-python==4.9.0.80
torch==2.3.1
ultralytics==8.2.37
numpy==1.26.4
Pillow==10.3.0
timm==0.9.16
torchvision==0.18.1 # Added for torchvision.models and transforms
